+++ src/components/common/ClientsListItem.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { StyleSheet } from 'react-native';
 import { List, Surface, TouchableRipple } from 'react-native-paper';
 
-import { navigationShape } from './propTypeShapes';
+import { navigationShape } from '../propTypeShapes';
 
 const styles = StyleSheet.create({
   listItem: {
